sha512_mb: extend max size to over 4GB

+++ sha512_mb/sha512_ctx_avx.c
@@ -36,7 +36,7 @@
 #endif
 
 static inline void hash_init_digest(SHA512_WORD_T * digest);
-static inline uint32_t hash_pad(uint8_t padblock[SHA512_BLOCK_SIZE * 2], uint32_t total_len);
+static inline uint32_t hash_pad(uint8_t padblock[SHA512_BLOCK_SIZE * 2], uint64_t total_len);
 static SHA512_HASH_CTX *sha512_ctx_mgr_resubmit(SHA512_HASH_CTX_MGR * mgr,
 						SHA512_HASH_CTX * ctx);
 
@@ -219,9 +219,9 @@ static inline void hash_init_digest(SHA512_WORD_T * digest)
 	memcpy_fixedlen(digest, hash_initial_digest, sizeof(hash_initial_digest));
 }
 
-static inline uint32_t hash_pad(uint8_t padblock[SHA512_BLOCK_SIZE * 2], uint32_t total_len)
+static inline uint32_t hash_pad(uint8_t padblock[SHA512_BLOCK_SIZE * 2], uint64_t total_len)
 {
-	uint32_t i = total_len & (SHA512_BLOCK_SIZE - 1);
+	uint32_t i = (uint32_t) (total_len & (SHA512_BLOCK_SIZE - 1));
 
 	memclr_fixedlen(&padblock[i], SHA512_BLOCK_SIZE);
 	padblock[i] = 0x80;
